+++ ChangeLog.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 
 ## Version 1.24
 
+- FIX : Disable board refresh will user grad an element *09/01/2023* - 1.24.6
 - FIX : **IMPORTANT Cache fix** *09/01/2023* - 1.24.5
 - FIX : Live update fix missing display return after update *08/01/2023* - 1.24.4
 - FIX : Can't split card from kanban when overspend *07/01/2023* - 1.24.3