+++ ChangeLog.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+
+
+## Version 1.27
+
 - NEW : remove clone function *04/01/2022* - 1.27.0
 - NEW : Ajout extrafield sur catégorie pour faire apparaitre le tag catégorie *04/01/2022* - 1.26.0  
   -> lors du préremplissage des notes de saisie de temps sur les tâches